refactor(schedule-details): split chart scales into per-function helpers
Move the combined scales module into helpers/, one function per file, to
match the project's established helper convention (e.g.
workflow-history-timeline-chart/helpers/). Rename resolveRunsChartTimeDomain
to resolveChartTimeWindow and drop the "Runs" prefix across these functions
since none of them read run/workflow data; "Domain" was also renamed to
"TimeWindow" to avoid confusion with Cadence's own domain concept. Tests move
to .test.ts since none of this logic touches Node.js APIs, and a coverage
gap in the time-window branch logic is closed with two additional cases.
